+++ spotbugs/src/main/java/edu/umd/cs/findbugs/sarif/Rule.java
@@ -42,10 +42,12 @@ final class Rule {
 
     JSONObject toJSONObject() {
         JSONObject messageStrings = new JSONObject().put("default", new JSONObject().put("text", defaultText));
-        JSONObject result = new JSONObject().put("id", id).put("shortDescription", new JSONObject().put("text", shortDescription)).put(
-                "fullDescription", new JSONObject().put("markdown", fullDescription)).put(
-                        "messageStrings",
-                        messageStrings).putOpt("helpUri", helpUri);
+        // TODO put 'fullDescription' with both of text and markdown representations
+        JSONObject result = new JSONObject()
+                .put("id", id)
+                .put("shortDescription", new JSONObject().put("text", shortDescription))
+                .put("messageStrings", messageStrings)
+                .putOpt("helpUri", helpUri);
         if (!tags.isEmpty()) {
             JSONObject propertyBag = new JSONObject().put("tags", new JSONArray(tags));
             result.put("properties", propertyBag);
@@ -54,7 +56,7 @@ final class Rule {
     }
 
     @NonNull
-    static Rule fromBugPattern(BugPattern bugPattern) {
+    static Rule fromBugPattern(BugPattern bugPattern, String formattedMessage) {
         URI helpUri = bugPattern.getUri().orElse(null);
 
         String category = bugPattern.getCategory();
@@ -65,7 +67,7 @@ final class Rule {
             tags = Collections.singletonList(category);
         }
 
-        return new Rule(bugPattern.getType(), bugPattern.getShortDescription(), bugPattern.getDetailText(), bugPattern.getLongDescription(), helpUri,
+        return new Rule(bugPattern.getType(), bugPattern.getShortDescription(), bugPattern.getDetailText(), formattedMessage, helpUri,
                 tags);
     }
 }